added check node version to scripts for windows

+++ mina/monitor/client/install.ps1
@@ -1,9 +1,27 @@
 param($b, $t)
 
-cls
 Write-Host -ForegroundColor green "Welcome to Mina Monitor Client installer!"
 Write-Host -ForegroundColor white "Copyright 2021 by Serhii Pimenov <serhii@pimenov.com.ua>"
+Read-Host "If you are ready, press [Enter] key to start or Ctrl+C to stop..."
 Write-Host
+
+Write-Host -ForegroundColor yellow "Check NodeJS..." -NoNewLine
+
+$node = (node -v | Select -First 1).replace("v", "")
+
+if (!$node) {
+    Write-Host -ForegroundColor red "Error!" -NoNewLine; Write-Host " NodeJS not installed! Please install NodeJS v14+ and try again."
+    exit
+}
+
+$nodeVer = $node.Split(".")
+
+if ($nodeVer[0] -lt 14) {
+    Write-Host -ForegroundColor red "Error!" -NoNewLine; Write-Host " Wrong NodeJS version! Please install NodeJS v14+ and try again."
+    exit
+}
+
+Write-Host -ForegroundColor green "OK"
 
 $br =  $b
 if (!$br) {
@@ -20,7 +38,9 @@ Write-Host -ForegroundColor yellow $br -NoNewLine
 Write-Host " branch into a folder " -NoNewLine
 Write-Host -ForegroundColor yellow $trg
 
-mkdir -p $trg
+if (-not(Test-Path -Path $trg)) {
+    mkdir -p $trg
+}
 cd $trg
 
 $sourcesUrl = "https://github.com/olton/mina-node-monitor/tarball/$br"
@@ -34,14 +54,18 @@ Write-Host "Sources in $url"
 
 tar --strip-components=2 -xf _.tar.gz "$url/client"
 
-Move-Item -Path config.example.json -Destination config.json
-
-Write-Host "Deleting temporary files..."
-
-Remove-Item -Path _.tar.gz
+if (-not(Test-Path -Path config.json)) {
+    Write-Host "Creating config file..." -NoNewLine
+    Move-Item -Path config.example.json -Destination config.json
+    Write-Host -ForegroundColor green "OK"
+}
 
 Write-Host "Installing dependencies..."
 npm install --silent
+
+Write-Host "Deleting temporary files..." -NoNewLine
+Remove-Item -Path _.tar.gz
+Write-Host -ForegroundColor green "OK"
 
 Write-Host
 Write-Host "Mina Monitor Client successfully installed."
